Publish version 0.0.1 of Etch-a-Sketch
Create the setup of the files needed for this project. Is currently on
step 2 of the instructions on the Odin Project website. currently trying
to create a grid for the etch-a-sketch. Having trouble using flexbox and
implementing a nested loop (because I can't do flex-basis). Looking into
using CSS Grid to solve the issue.