Publish version 0.0.2 of Odinproject Etch-a-Sketch
Commit code for the basic version of the Etch-a-Sketch project. As of
now, players can manually input the tile size in to the code, and if it
is an square number, the appropriate number of tiles will show on the
computer. Currently, when the user hovers over the tiles, the CSS backgr
ound property changes to blue. Also added a clear button to reset the
game. Looking into a drag and hover effect for the tiles, as well as
a GUI input into the number of tiles as well as a color picker option.