Publish version 0.0.3 (alpha) of Etch-a-Sketch
Commit code for the alpha version of the Etch-a-Sketch application for
the Odin Project. Added a slider functionality that allows users to
adjust the amount of tiles in the etch-a-sketch app. Right now the app
just needs a few touch ups. Drag and draw functionality, initial value
of the grid are two things that come to mind, as well as some extra
styling of the body. But importantly I think I need to create some
dynamic sizing of the tiles so I can get bigger sizes into the app.